requirements.txt
absl-py==0.2.0
actionlib==1.11.13
angles==1.9.11
astor==0.6.2
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
bondpy==1.8.1
camera-calibration==1.12.22
camera-calibration-parsers==1.11.13
catkin==0.7.11
catkin-pkg==0.4.1
controller-manager==0.13.1
controller-manager-msgs==0.13.1
cv-bridge==1.12.7
cycler==0.10.0
decorator==4.3.0
diagnostic-analysis==1.9.2
diagnostic-common-diagnostics==1.9.2
diagnostic-updater==1.9.2
docutils==0.14
dynamic-reconfigure==1.5.49
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
gazebo-plugins==2.5.14
gazebo-ros==2.5.14
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.10
gennodejs==2.0.1
genpy==0.6.7
grpcio==1.11.0
html5lib==0.9999999
image-geometry==1.12.7
interactive-markers==1.11.3
kiwisolver==1.0.1
laser-geometry==1.6.4
Markdown==2.6.11
matplotlib==2.2.2
message-filters==1.12.13
mock==2.0.0
moveit-commander==0.9.11
moveit-ros-planning-interface==0.9.11
moveit-ros-visualization==0.9.11
netifaces==0.10.6
networkx==2.1
numpy==1.14.2
pbr==4.0.2
Pillow==5.1.0
pluginlib==1.11.3
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.2
python-qt-binding==0.3.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
qt-dotgraph==0.3.8
qt-gui==0.3.8
qt-gui-cpp==0.3.8
qt-gui-py-common==0.3.8
resource-retriever==1.12.4
rosbag==1.12.13
rosboost-cfg==1.14.3
rosclean==1.14.3
roscreate==1.14.3
rosgraph==1.12.13
roslaunch==1.12.13
roslib==1.14.3
roslint==0.11.0
roslz4==1.12.13
rosmake==1.14.3
rosmaster==1.12.13
rosmsg==1.12.13
rosnode==1.12.13
rosparam==1.12.13
rospkg==1.1.4
rospy==1.12.13
rosservice==1.12.13
rostest==1.12.13
rostopic==1.12.13
rosunit==1.14.3
roswtf==1.12.13
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.8
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.11
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.9
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.7
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.15
scikit-image==0.13.1
scipy==1.0.1
sensor-msgs==1.12.5
six==1.11.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.1
srdfdom==0.4.2
subprocess32==3.2.7
tensorboard==1.7.0
tensorflow==1.7.0
tensorlayer==1.8.4
termcolor==1.1.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.17
tf2-kdl==0.5.17
tf2-py==0.5.17
tf2-ros==0.5.17
topic-tools==1.12.13
trac-ik-python==1.4.11
urdfdom-py==0.3.3
Werkzeug==0.14.1
xacro==1.11.3
